refactor: conform the dataloaders package to Principle VI
Reword the three module docstrings to one imperative line and merge the
_schema helpers into _base so the base module imports no sibling, per the
self-contained-base rule. The schema builders were never public, so the
public API and behaviour are unchanged.
